refactor: extract utils function from useProjectsSellOrders
